pyproject.toml
[project]
name = "job-scraper"
version = "0.1.0"
description = "Scraps jobs form a website based on keywords"
authors = [
    {name = "Harpreet Dhillon"}
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "fuzzywuzzy (>=0.18.0,<0.19.0)",
    "selenium (>=4.30.0,<5.0.0)",
    "rapidfuzz (>=3.12.2,<4.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "pydantic (>=2.11.2,<3.0.0)",
    "tls-client (>=1.0.1,<2.0.0)",
    "markdownify (>=1.1.0,<2.0.0)",
    "regex (>=2024.11.6,<2025.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

